[refactor]: 피드 프로필 누르면 화면이동하도록 수정 (#101)

+++ app/src/main/java/com/texthip/thip/data/model/feed/response/AllFeedResponse.kt
@@ -14,7 +14,7 @@ data class AllFeedResponse(
 @Serializable
 data class AllFeedItem(
     @SerialName("feedId") val feedId: Int,
-    @SerialName("creatorId") val creatorId: Int,
+    @SerialName("creatorId") val creatorId: Long,
     @SerialName("creatorNickname") val creatorNickname: String,
     @SerialName("creatorProfileImageUrl") val creatorProfileImageUrl: String?,
     @SerialName("aliasName") val aliasName: String,
